requirements.txt
# ── LLM / Agent orchestration ─────────────────────────────────────────────────
crewai==1.10.1
google-generativeai==0.8.6
langchain==1.2.10
langchain-core==1.2.19
langchain-community==0.4.1
langchain-google-genai==4.2.1

# ── RAG / Vector store ────────────────────────────────────────────────────────
chromadb==1.5.5
sentence-transformers==5.3.0

# ── Data processing ───────────────────────────────────────────────────────────
pandas==2.3.1
numpy==2.3.2
pdfplumber==0.11.9
openpyxl==3.1.5

# ── Visualisation ─────────────────────────────────────────────────────────────
plotly==6.3.0
matplotlib==3.10.5
kaleido==0.2.1          # 0.2.x uses bundled binary (no Chrome needed); 1.x requires Chrome

# ── UI ────────────────────────────────────────────────────────────────────────
gradio==6.9.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.1.1
huggingface_hub>=0.23.0
requests==2.32.5
httpx==0.28.1
uvicorn==0.41.0
